move registration summary to volunteer
